Finish refactoring ChatToolInvocationPart Got sidetracked trying to make Copilot do this and forgot the last class.
