+++ src/vs/workbench/contrib/chat/browser/chatContentParts/toolInvocationParts/chatToolInvocationPart.ts
@@ -4,28 +4,25 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as dom from '../../../../../../base/browser/dom.js';
-import { Codicon } from '../../../../../../base/common/codicons.js';
 import { Emitter } from '../../../../../../base/common/event.js';
-import { IMarkdownString, MarkdownString } from '../../../../../../base/common/htmlContent.js';
 import { Disposable, DisposableStore, IDisposable } from '../../../../../../base/common/lifecycle.js';
-import { autorunWithStore } from '../../../../../../base/common/observable.js';
 import { MarkdownRenderer } from '../../../../../../editor/browser/widget/markdownRenderer/browser/markdownRenderer.js';
 import { IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
-import { IChatProgressMessage, IChatToolInvocation, IChatToolInvocationSerialized } from '../../../common/chatService.js';
+import { IChatToolInvocation, IChatToolInvocationSerialized } from '../../../common/chatService.js';
 import { IChatRendererContent } from '../../../common/chatViewModel.js';
 import { CodeBlockModelCollection } from '../../../common/codeBlockModelCollection.js';
 import { isToolResultInputOutputDetails } from '../../../common/languageModelToolsService.js';
 import { ChatTreeItem, IChatCodeBlockInfo } from '../../chat.js';
 import { IChatContentPart, IChatContentPartRenderContext } from '../chatContentParts.js';
 import { EditorPool } from '../chatMarkdownContentPart.js';
-import { ChatProgressContentPart } from '../chatProgressContentPart.js';
 import { CollapsibleListPool } from '../chatReferencesContentPart.js';
+import { ChatInputOutputMarkdownProgressPart } from './chatInputOutputMarkdownProgressPart.js';
+import { ChatResultListSubPart } from './chatResultListSubPart.js';
 import { ChatTerminalMarkdownProgressPart } from './chatTerminalMarkdownProgressPart.js';
 import { TerminalConfirmationWidgetSubPart } from './chatTerminalToolSubPart.js';
 import { ToolConfirmationSubPart } from './chatToolConfirmationSubPart.js';
-import { ChatResultListSubPart } from './chatResultListSubPart.js';
-import { ChatInputOutputMarkdownProgressPart } from './chatInputOutputMarkdownProgressPart.js';
 import { BaseChatToolInvocationSubPart } from './chatToolInvocationSubPart.js';
+import { ChatToolProgressSubPart } from './chatToolProgressPart.js';
 
 export class ChatToolInvocationPart extends Disposable implements IChatContentPart {
 	public readonly domNode: HTMLElement;
@@ -41,7 +38,7 @@ export class ChatToolInvocationPart extends Disposable implements IChatContentPa
 		return this.subPart?.codeblocksPartId;
 	}
 
-	private subPart!: ChatToolInvocationSubPart | BaseChatToolInvocationSubPart;
+	private subPart!: BaseChatToolInvocationSubPart;
 
 	constructor(
 		private readonly toolInvocation: IChatToolInvocation | IChatToolInvocationSerialized,
@@ -80,7 +77,7 @@ export class ChatToolInvocationPart extends Disposable implements IChatContentPa
 		render();
 	}
 
-	createToolInvocationSubPart(): ChatToolInvocationSubPart | BaseChatToolInvocationSubPart {
+	createToolInvocationSubPart(): BaseChatToolInvocationSubPart {
 		if (this.toolInvocation.kind === 'toolInvocation' && this.toolInvocation.confirmationMessages) {
 			if (this.toolInvocation.toolSpecificData?.kind === 'terminal') {
 				return this.instantiationService.createInstance(TerminalConfirmationWidgetSubPart, this.toolInvocation, this.toolInvocation.toolSpecificData, this.context, this.renderer, this.editorPool, this.currentWidthDelegate, this.codeBlockStartIndex);
@@ -127,7 +124,7 @@ export class ChatToolInvocationPart extends Disposable implements IChatContentPa
 			);
 		}
 
-		return this.instantiationService.createInstance(ChatToolInvocationSubPart, this.toolInvocation, this.context, this.renderer);
+		return this.instantiationService.createInstance(ChatToolProgressSubPart, this.toolInvocation, this.context, this.renderer);
 	}
 
 	hasSameContent(other: IChatRendererContent, followingContent: IChatRendererContent[], element: ChatTreeItem): boolean {
@@ -136,73 +133,5 @@ export class ChatToolInvocationPart extends Disposable implements IChatContentPa
 
 	addDisposable(disposable: IDisposable): void {
 		this._register(disposable);
-	}
-}
-
-class ChatToolInvocationSubPart extends Disposable {
-	private static idPool = 0;
-
-	private readonly _codeblocksPartId = 'tool-' + (ChatToolInvocationSubPart.idPool++);
-
-	public readonly domNode: HTMLElement;
-
-	private _onNeedsRerender = this._register(new Emitter<void>());
-	public readonly onNeedsRerender = this._onNeedsRerender.event;
-
-	private _onDidChangeHeight = this._register(new Emitter<void>());
-	public readonly onDidChangeHeight = this._onDidChangeHeight.event;
-
-	private _codeblocks: IChatCodeBlockInfo[] = [];
-	public get codeblocks(): IChatCodeBlockInfo[] {
-		return this._codeblocks;
-	}
-
-	public get codeblocksPartId(): string {
-		return this._codeblocksPartId;
-	}
-
-	constructor(
-		private readonly toolInvocation: IChatToolInvocation | IChatToolInvocationSerialized,
-		private readonly context: IChatContentPartRenderContext,
-		private readonly renderer: MarkdownRenderer,
-		@IInstantiationService private readonly instantiationService: IInstantiationService,
-	) {
-		super();
-
-		this.domNode = this.createProgressPart();
-	}
-
-	private createProgressPart(): HTMLElement {
-		if (this.toolInvocation.isComplete && this.toolInvocation.isConfirmed !== false && this.toolInvocation.pastTenseMessage) {
-			const part = this.renderProgressContent(this.toolInvocation.pastTenseMessage);
-			this._register(part);
-			return part.domNode;
-		} else {
-			const container = document.createElement('div');
-			const progressObservable = this.toolInvocation.kind === 'toolInvocation' ? this.toolInvocation.progress : undefined;
-			this._register(autorunWithStore((reader, store) => {
-				const progress = progressObservable?.read(reader);
-				const part = store.add(this.renderProgressContent(progress?.message || this.toolInvocation.invocationMessage));
-				dom.reset(container, part.domNode);
-			}));
-			return container;
-		}
-	}
-
-	private renderProgressContent(content: IMarkdownString | string) {
-		if (typeof content === 'string') {
-			content = new MarkdownString().appendText(content);
-		}
-
-		const progressMessage: IChatProgressMessage = {
-			kind: 'progressMessage',
-			content
-		};
-
-		const iconOverride = !this.toolInvocation.isConfirmed ?
-			Codicon.error :
-			this.toolInvocation.isComplete ?
-				Codicon.check : undefined;
-		return this.instantiationService.createInstance(ChatProgressContentPart, progressMessage, this.renderer, this.context, undefined, true, iconOverride);
 	}
 }
