+++ src/vs/workbench/contrib/chat/browser/chatContentParts/toolInvocationParts/chatToolProgressPart.ts
@@ -1,0 +1,67 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as dom from '../../../../../../base/browser/dom.js';
+import { Codicon } from '../../../../../../base/common/codicons.js';
+import { IMarkdownString, MarkdownString } from '../../../../../../base/common/htmlContent.js';
+import { autorunWithStore } from '../../../../../../base/common/observable.js';
+import { MarkdownRenderer } from '../../../../../../editor/browser/widget/markdownRenderer/browser/markdownRenderer.js';
+import { IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
+import { IChatToolInvocation, IChatToolInvocationSerialized, IChatProgressMessage } from '../../../common/chatService.js';
+import { IChatCodeBlockInfo } from '../../chat.js';
+import { IChatContentPartRenderContext } from '../chatContentParts.js';
+import { ChatProgressContentPart } from '../chatProgressContentPart.js';
+import { BaseChatToolInvocationSubPart } from './chatToolInvocationSubPart.js';
+
+export class ChatToolProgressSubPart extends BaseChatToolInvocationSubPart {
+	public readonly domNode: HTMLElement;
+
+	public override readonly codeblocks: IChatCodeBlockInfo[] = [];
+
+	constructor(
+		private readonly toolInvocation: IChatToolInvocation | IChatToolInvocationSerialized,
+		private readonly context: IChatContentPartRenderContext,
+		private readonly renderer: MarkdownRenderer,
+		@IInstantiationService private readonly instantiationService: IInstantiationService,
+	) {
+		super(toolInvocation);
+
+		this.domNode = this.createProgressPart();
+	}
+
+	private createProgressPart(): HTMLElement {
+		if (this.toolInvocation.isComplete && this.toolInvocation.isConfirmed !== false && this.toolInvocation.pastTenseMessage) {
+			const part = this.renderProgressContent(this.toolInvocation.pastTenseMessage);
+			this._register(part);
+			return part.domNode;
+		} else {
+			const container = document.createElement('div');
+			const progressObservable = this.toolInvocation.kind === 'toolInvocation' ? this.toolInvocation.progress : undefined;
+			this._register(autorunWithStore((reader, store) => {
+				const progress = progressObservable?.read(reader);
+				const part = store.add(this.renderProgressContent(progress?.message || this.toolInvocation.invocationMessage));
+				dom.reset(container, part.domNode);
+			}));
+			return container;
+		}
+	}
+
+	private renderProgressContent(content: IMarkdownString | string) {
+		if (typeof content === 'string') {
+			content = new MarkdownString().appendText(content);
+		}
+
+		const progressMessage: IChatProgressMessage = {
+			kind: 'progressMessage',
+			content
+		};
+
+		const iconOverride = !this.toolInvocation.isConfirmed ?
+			Codicon.error :
+			this.toolInvocation.isComplete ?
+				Codicon.check : undefined;
+		return this.instantiationService.createInstance(ChatProgressContentPart, progressMessage, this.renderer, this.context, undefined, true, iconOverride);
+	}
+}
